--- 2545.go
@@ -1,42 +1,0 @@
-package dev
-
-import (
-	"megaCrawler/crawlers"
-	"megaCrawler/extractors"
-	"strings"
-	"time"
-
-	"github.com/gocolly/colly/v2"
-)
-
-func init() {
-	engine := crawlers.Register("2545", "WEEK", "https://www.seafdec.org.ph")
-	engine.SetTimeout(60 * time.Second)
-
-	engine.SetStartingURLs([]string{"https://www.seafdec.org.ph/post-sitemap.xml"})
-
-	extractorConfig := extractors.Config{
-		Author:       true,
-		Image:        false,
-		Language:     true,
-		PublishDate:  true,
-		Tags:         true,
-		Text:         false,
-		Title:        true,
-		TextLanguage: "",
-	}
-
-	extractorConfig.Apply(engine)
-
-	engine.OnXML("//loc", func(element *colly.XMLElement, ctx *crawlers.Context) {
-		if strings.Contains(element.Text, "post-sitemap") {
-			engine.Visit(element.Text, crawlers.Index)
-		} else if !strings.Contains(element.Text, ".xml") {
-			engine.Visit(element.Text, crawlers.News)
-		}
-	})
-
-	engine.OnHTML("#the-post > div > div.entry > p", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Content += element.Text
-	})
-}
