--- 2603.go
@@ -1,51 +1,0 @@
-package dev
-
-import (
-	"megaCrawler/crawlers"
-	"megaCrawler/extractors"
-
-	"github.com/gocolly/colly/v2"
-)
-
-func init() {
-	engine := crawlers.Register("2603", "FAO in the Philippines", "https://www.fao.org/philippines/en/")
-
-	engine.SetStartingURLs([]string{"https://www.fao.org/philippines/news/en/"})
-
-	extractorConfig := extractors.Config{
-		Author:       true,
-		Image:        false,
-		Language:     true,
-		PublishDate:  true,
-		Tags:         true,
-		Text:         false,
-		Title:        true,
-		TextLanguage: "",
-	}
-
-	extractorConfig.Apply(engine)
-	engine.OnHTML("div.list-block > div.list-title > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		url, err := element.Request.URL.Parse(element.Attr("href"))
-
-		if err != nil {
-			crawlers.Sugar.Error(err.Error())
-			return
-		}
-		engine.Visit(url.String(), crawlers.News)
-	})
-
-	engine.OnHTML("#c245032 > div > div.tx-dynalist-pi1-pagination > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		url, err := element.Request.URL.Parse(element.Attr("href"))
-
-		if err != nil {
-			crawlers.Sugar.Error(err.Error())
-			return
-		}
-		engine.Visit(url.String(), crawlers.Index)
-	})
-
-	engine.OnHTML("#c245038 > div > div > div.news-item > p", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Content += element.Text
-	})
-
-}
