--- 2640.go
@@ -1,35 +1,0 @@
-package dev
-
-import (
-	"megaCrawler/crawlers"
-	"megaCrawler/extractors"
-
-	"github.com/gocolly/colly/v2"
-)
-
-func init() {
-	engine := crawlers.Register("2640", "Capital University", "https://www.cu.edu.ph")
-
-	engine.SetStartingURLs([]string{"https://www.cu.edu.ph/news-articles/"})
-
-	extractorConfig := extractors.Config{
-		Author:       true,
-		Image:        false,
-		Language:     true,
-		PublishDate:  true,
-		Tags:         true,
-		Text:         false,
-		Title:        true,
-		TextLanguage: "",
-	}
-
-	extractorConfig.Apply(engine)
-	engine.OnHTML(" div > div > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		engine.Visit(element.Attr("href"), crawlers.News)
-	})
-
-	engine.OnHTML(" div > div.entry-content.single-page > p", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Content += element.Text
-	})
-
-}
