--- 2785.go
@@ -1,42 +1,0 @@
-package dev
-
-import (
-	"megaCrawler/crawlers"
-	"megaCrawler/extractors"
-
-	"github.com/gocolly/colly/v2"
-)
-
-func init() {
-	engine := crawlers.Register("2785", "NHCP", "https://nhcp.gov.ph")
-
-	engine.SetStartingURLs([]string{"https://nhcp.gov.ph/news_and_updates/"})
-
-	extractorConfig := extractors.Config{
-		Author:       false,
-		Image:        false,
-		Language:     true,
-		PublishDate:  false,
-		Tags:         true,
-		Text:         false,
-		Title:        true,
-		TextLanguage: "",
-	}
-
-	extractorConfig.Apply(engine)
-	engine.OnHTML("div.arihive-col-two > h2 > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		engine.Visit(element.Attr("href"), crawlers.News)
-	})
-
-	engine.OnHTML("div.pagination > div > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		engine.Visit(element.Attr("href"), crawlers.Index)
-	})
-
-	engine.OnHTML("div > p", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Content += element.Text
-	})
-
-	engine.OnHTML("div > span.posted-on > a > time", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.PublicationTime += element.Text
-	})
-}
