+++ plugins/production/2606.go
@@ -1,0 +1,57 @@
+package dev
+
+import (
+	"megaCrawler/crawlers"
+	"megaCrawler/extractors"
+	"time"
+
+	"github.com/gocolly/colly/v2"
+)
+
+func init() {
+	engine := crawlers.Register("2606", "GEF", "https://www.thegef.org")
+	engine.SetTimeout(60 * time.Second)
+
+	engine.SetStartingURLs([]string{"https://www.thegef.org/newsroom/news"})
+
+	extractorConfig := extractors.Config{
+		Author:       true,
+		Image:        false,
+		Language:     true,
+		PublishDate:  false,
+		Tags:         true,
+		Text:         false,
+		Title:        true,
+		TextLanguage: "",
+	}
+
+	extractorConfig.Apply(engine)
+	engine.OnHTML("div > div.wbcard__text > h3 > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		url, err := element.Request.URL.Parse(element.Attr("href"))
+
+		if err != nil {
+			crawlers.Sugar.Error(err.Error())
+			return
+		}
+		engine.Visit(url.String(), crawlers.News)
+	})
+
+	engine.OnHTML("#block-geftheme-content > div > div > div > nav > ul > li > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		url, err := element.Request.URL.Parse(element.Attr("href"))
+
+		if err != nil {
+			crawlers.Sugar.Error(err.Error())
+			return
+		}
+		engine.Visit(url.String(), crawlers.Index)
+	})
+
+	engine.OnHTML("div.clearfix.text-formatted.field.field--name-body.field--type-text-with-summary.field--label-hidden.field__item > p", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Content += element.Text
+	})
+
+	engine.OnHTML("div.field.field--name-field-date.field--type-datetime.field--label-hidden.field__item > time", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.PublicationTime += element.Text
+	})
+
+}
